tracer.py: Moved TracerMObject to its own module.
